test: increase compliance with @babel/parser's test suite
Previously, jam's babel test runner
used `std.mem.eql` to compare two JSON-serialized ASTs.
This was obviously flawed, because then things the order of
keys in an object start to matter.

With this commit, we use a deep equality checker instead.

This commit also fixes a UB caused by `Tree.getNode`
returning the address of a stack local.