chore(benchmark): Add benchmark suite for the parser
Currently, the benchmark script (benchmarl/bench-parser.zig) just
stores the benchmark result in './parser-bench.json' file.

Ideally, it should just print the time taken to `stdout` and not write
to any files. This way, we can have a CI job that runs the benchmark
on 'HEAD', and then again on the main branch, and reports the perf
delta in a comment on codeberg/gh.